requirements.txt
pytest==7.4.0
PyYAML==6.0.1
mysql-connector-python==8.1.0
requests==2.31.0
allure-pytest==2.13.2
pytest-html==4.1.1